Kunskapskontroll 2/requirements.txt
pandas
matplotlib
python-dateutil
statsmodels